+++ src/Components/index.ts
@@ -1,3 +1,3 @@
-export { default as IPGComponent } from './IPGComponent';
+//export { default as IPGClassComponent } from './IPGClassComponent';
 export { default as IPGStage } from './IPGStage';
-export { default as  IPGFunctionalComponent } from './IPGFunctionalComponent';
+export { default as  IPGComponent } from './IPGComponent';
